requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0

# AI/LLM
anthropic==0.42.0
openai==1.58.1
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
langgraph-checkpoint-sqlite>=2.0.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6

# Redis & Task Queue
redis==5.2.1
arq==0.26.1

# Communication Channels
twilio==9.4.0
resend==2.6.0
websockets==14.1
tweepy==4.14.0

# Utilities
httpx==0.28.1
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Document Parsing (RAG)
pypdf==5.1.0
python-docx==1.1.2

# Google APIs (Calendar OAuth)
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Token Encryption
cryptography==42.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
factory-boy==3.3.1
faker==33.1.0
aiosqlite==0.20.0
respx==0.22.0
